ProAct_single&batch/requirements.txt
uvicorn==0.22.0
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0
imbalanced-learn>=0.11.0
requests>=2.31.0
streamlit>=1.29.0
matplotlib>=3.8.0
seaborn>=0.13.0
python-multipart>=0.0.6
google-generativeai==0.3.1
python-dotenv==1.0.0
